Expose remaining_time and ledge_grabs as agent observations
Completes the agent-observation work begun in the stocks PR. Both fields are
computed at parse time and persisted in the NamedTuple schema, avoiding any
changes to the data-loading pipeline.

- slippi_ai/types.py: new ledge_grabs: np.uint16 on Player and Nana (after
  stocks), and remaining_time: np.float32 on Game (after stage).
- slippi_db/parse_peppi.py: populates ledge_grabs via the existing
  cumulative_ledge_grabs helper and computes remaining_time from
  peppi_game.start.timer (seconds) normalized over the frame index.
- slippi_db/parse_libmelee.py: mirrors ledge_grabs tracking per port across
  frames and derives remaining_time from game.frame using a timer_seconds
  parameter (default 480s, the tournament timer).
- slippi_ai/embed.py: adds optional embed_ledge_grabs (one-hot, 50 buckets,
  CLAMP) on the player embedding and embed_remaining_time (float in [0, 1])
  on the game embedding, both gated by new config flags and defaulting on.
- slippi_ai/custom/tournament_rules.py: cumulative_ledge_grabs now returns
  uint16 to match the Player field dtype and avoid overflow in pathological
  timeout games.
- Tests cover the dtype change and a 300-grab non-overflow scenario.

+++ slippi_ai/custom/tournament_rules.py
@@ -50,7 +50,7 @@ def cumulative_ledge_grabs(action_array: np.ndarray) -> np.ndarray:
   Output has the same length as the input and is a running count of
   transitions into EDGE_CATCHING up to and including each frame.
   """
-  return np.cumsum(_entry_mask(action_array)).astype(np.uint8)
+  return np.cumsum(_entry_mask(action_array)).astype(np.uint16)
 
 
 def game_ended_by_timeout(last_frame: int, timer_seconds: int) -> bool:
